Update viewer_projection.cpp
Ensures the proper tooltip is displayed when no layers are loaded. In this situation the active_layers variable also = 0, so the "no layers active" tooltip is displayed. This is obviously useless - the user should be told that there are no projection layers in the set full stop - active or not.